requirements.txt
Django==1.10.1
mysqlclient==1.3.7
Pillow==3.4.0
pylabels==1.2.1
PyMySQL==0.7.9
reportlab==3.3.0
